LayoutTests/userscripts/user-script-image-document-expected.txt

CONSOLE MESSAGE: line 1: Start Script
CONSOLE MESSAGE: line 1: End Script
CONSOLE MESSAGE: line 1: Start Script
CONSOLE MESSAGE: line 1: End Script
